init opencl done

+++ sources/math_utils/math_utils.c
@@ -1,6 +1,6 @@
 #include "math_utils.h"
 
-int 				in_range_inclusive(double number, double min, double max)
+int 				in_range_inclusive(float number, float min, float max)
 {
 	if (number >= min && number <= max)
 		return (TRUE);
